\name{MetabolAnalyze-package}
\Rdversion{1.1}
\alias{MetabolAnalyze-package}
\alias{MetabolAnalyze}
\docType{package}
\title{
Probabilistic latent variable models for metabolomic data.
}
\description{
Fits probabilistic principal components analysis (PPCA), probabilistic principal components and covariates analysis (PPCCA) and mixtures of probabilistic principal component analysis (MPPCA) models to metabolomic spectral data. Estimates of the uncertainty associated with the model parameter estimates are provided.
}
\details{
\tabular{ll}{
Package: \tab MetabolAnalyze\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-05-12\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}
\author{
Nyamundanda Gift, Isobel Claire Gormley and Lorraine Brennan.

Claire Gormley <claire.gormley@ucd.ie>
}
\references{
Nyamundanda, G., Gormley, I.C. and Brennan, L. (2010) Probabilistic principal components analysis for metabolomic data. Technical Report. University College Dublin.
}

